### 2019-07-04 #16 @whedon
[PRE REVIEW]: Locality-Aware Message Passing with ArrayChannels.jl
**Submitting author:** @rohanmclure (<a href="http://orcid.org/0000-0002-7018-5501">Rohan McLure</a>)
**Repository:** <a href="https://github.com/rohanmclure/ArrayChannels.jl" target ="_blank">https://github.com/rohanmclure/ArrayChannels.jl</a>

Thanks for submitting your paper to JuliaCon @rohanmclure. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@rohanmclure if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
