### 2019-07-04 e207e02
Specify targets with ArrayChannel reference
