+++ src/precompile.jl
@@ -1,6 +1,6 @@
 precompile(Tuple{ArrayChannels.get_arraychannel, Distributed.RRID})
 precompile(Tuple{ArrayChannels.ac_get_from, Distributed.RRID})
-precompile(Tuple{ArrayChannels.put!, ArrayChannel, Int64, Union{RRID,Nothing}})
+precompile(Tuple{ArrayChannels.put!, ArrayChannel, Int64, Union{ArrayChannel,RRID,Nothing}})
 precompile(Tuple{ArrayChannels.take!, ArrayChannel, Int64})
 precompile(Tuple{ArrayChannels.put!, Handshake, Int64})
 precompile(Tuple{ArrayChannels.take!, Handshake, Int64})
