### 2019-07-06 #16 @rohanmclure
Very happy to have @jpsamaroo review the submission. It would be good to hear recommendations on figure formatting. Thanks
