### 2019-07-06 #20 @whedon
[REVIEW]: Locality-Aware Message Passing with ArrayChannels.jl
**Submitting author:** @rohanmclure (<a href="http://orcid.org/0000-0002-7018-5501">Rohan McLure</a>)
**Repository:** <a href="https://github.com/rohanmclure/ArrayChannels.jl" target ="_blank">https://github.com/rohanmclure/ArrayChannels.jl</a>
