### 2019-07-07 102c0fc
Update Dockerfile and install instructions
