### 2019-07-07 #20 @rohanmclure
Have just updated build instructions, as well as provided a Binder and Dockerfile.
