### 2019-07-11 rohanmclure/ArrayChannels.jl#1 @jpsamaroo
Installation instructions in README are incorrect
The instructions in the README do a `git clone` of the repo, but fails to `cd` into the newly-cloned directory. Without this step, there's no way for Pkg to know where the cloned ArrayChannels.jl repo is.

I would first recommend changing the name of the repo in the first step of the instructions like so: `git clone git@github.com:rohanmclure/ArrayChannels.jl.git ArrayChannels` (so that the name of the folder is "ArrayChannels", per Julia conventions), and then changing the last step of the instructions from `dev .` to `dev ./ArrayChannels`. Following these steps will also cause Pkg to automatically add all listed dependencies, so you should be able to remove the step containing `add Distributed Serialization Sockets`.

EDIT: X-Ref https://github.com/JuliaCon/proceedings-review/issues/20
