### 2019-07-11 rohanmclure/ArrayChannels.jl#2 @jpsamaroo
Dockerfile and docker instructions
When building the Dockerfile, you'll notice numerous ~errors~ warnings from `apt` indicating that the its API is not stable; it would be a good idea to heed the warnings and switch from `apt` to `apt-get` (even if the latter is a bit slower) so that a change to the CLI API doesn't accidentally break our Dockerfile.

The docker instructions generally work, however the `docker run` command does not publish the port, so one can't access the Jupyter instance from outside the container. Adding `-p 0.0.0.0:8888:8888` before `arraychannels` in the command fixes this and makes the Jupyter instance accessible from the host or other networked computers.

EDIT: X-Ref https://github.com/JuliaCon/proceedings-review/issues/20
