### 2019-07-11 rohanmclure/ArrayChannels.jl#2 @jpsamaroo
Also, it seems as if the IJulia kernel is not installed? I don't have an option for a Julia kernel in Jupyter, only one exists for Python 3.
