### 2019-07-11 rohanmclure/ArrayChannels.jl#3 @jpsamaroo
Missing DOI references
There are a number of DOI entries missing in `paper/ref.bib`, such as for "
Julia: A fresh approach to numerical computing" and "Structured Parallel Programming: Patterns for Efficient Computation}".

X-Ref https://github.com/JuliaCon/proceedings-review/issues/20
