### 2019-07-11 rohanmclure/ArrayChannels.jl#4 @jpsamaroo
Add CI and/or describe how to run tests
X-Ref https://github.com/JuliaCon/proceedings-review/issues/20
