### 2019-07-12 02c768b
Use apt-get and IJulia
