+++ Dockerfile
@@ -13,6 +13,6 @@ USER jupyter
 WORKDIR /home/jupyter/arraychannels
 RUN julia -e "using Pkg; Pkg.add(\"IJulia\"); using IJulia"
 RUN julia -e "using Pkg; Pkg.activate(\".\"); Pkg.resolve()"
-RUN julia -e "using Pkg; Pkg.add(\"Distributed\"); Pkg.add(\"Serialization\"); Pkg.add(\"Sockets\"); Pkg.develop(PackageSpec(path=\".\"))"
+RUN julia -e "using Pkg; Pkg.add(\"Distributed\"); Pkg.add(\"Serialization\"); Pkg.add(\"Sockets\"); Pkg.develop(PackageSpec(path=\".\")); using ArrayChannels"
 
 EXPOSE 8888
