+++ Makefile
@@ -1,10 +1,19 @@
-CC = gcc 
-MPICC = mpicc 
+CC = gcc
+MPICC = mpicc
 
 clean:
-	
+
 ping_pong:
 	$(MPICC) example/ping_pong.c -o example/ping_pong
 
-reduce:
-	$(MPICC) example/reduce.c -o example/reduce
+reduce: prk
+	cd example && $(MPICC) reduce.c -o reduce
+
+transpose:
+	cd example && $(MPICC) transpose.c -o transpose
+
+stencil:
+	cd example && $(MPICC) stencil.c -o stencil
+
+prk:
+	# Do something or rather else
