### 2019-07-14 e722c41
Update Dockerfile subject to issue 2
