+++ paper/ref.bib
@@ -3,6 +3,8 @@
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
+  url = {https://doi.org/10.1137/141000671},
+  doi = {10.1137/141000671},
   number={1},
   pages={65--98},
   year={2017},
@@ -13,6 +15,8 @@
   author = {Michael A. McCool and Arch D. Robinson and James Reinders},
   title = {Structured Parallel Programming: Patterns for Efficient Computation},
   subtitle = {Patterns for efficient computation},
+  url = {https://doi.org/10.1016/C2011-0-04251-5},
+  doi = {10.1016/C2011-004251-5},
   year = {2012},
   publisher = {Morgan Kaufman}
 }
