### 2019-07-14 #20 @rohanmclure
Sorry about the wait in acting on those issues, I have been in transit. I have addressed the four issues, and will set up CircleCI / Travis soon for CI, but testing documentation has been provided in `README.md`.

The Binder link should now provide a Jupyter notebook with Julia 1.1.0 and Python 3 kernels available.
