### 2019-07-14 #20 @jpsamaroo
Great! Is it possible to create DOI references for the lecture and screencast at the end of the `ref.bib`?
