### 2019-07-14 rohanmclure/ArrayChannels.jl#4 @jpsamaroo
Once CI is added and working, please feel free to close this @rohanmclure
