### 2019-07-14 rohanmclure/ArrayChannels.jl#5 @jpsamaroo
Add CONTRIBUTING.md
Should contain details about: "clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

X-Ref https://github.com/JuliaCon/proceedings-review/issues/20
