### 2019-07-15 #20 @rohanmclure
> Great! Is it possible to create DOI references for the lecture and screencast at the end of the `ref.bib`?

I don't have any way to generate DOI links, however I can provide URLs for each of them.
