+++ paper/ref.bib
@@ -76,6 +76,7 @@
   author = {William Gropp},
   title = {Lecture 24: Buffering and Message Protocols},
   year = {2016},
+  url = {http://wgropp.cs.illinois.edu/courses/cs598-s16/lectures/lecture24.pdf},
   publisher = {University of Illinois, Chicago}
 }
 
@@ -83,5 +84,6 @@
   author = {Jeff Squyres},
   title = {Screencast: OMPI OpenFabrics Protocols},
   year = {2008},
+  url = {https://www.open-mpi.org/video/openfabrics/ompi-openfabrics-protocols-v1.2-2up.pdf},
   publisher = {CISCO}
 }
