### 2019-07-16 #20 @vchuravy
You can deposit them with Zenodo.
