### 2019-07-17 83b8cc4
Add DOI references
