+++ paper/ref.bib
@@ -77,6 +77,7 @@
   title = {Lecture 24: Buffering and Message Protocols},
   year = {2016},
   url = {http://wgropp.cs.illinois.edu/courses/cs598-s16/lectures/lecture24.pdf},
+  doi = {10.5281/zenodo.3339229},
   publisher = {University of Illinois, Chicago}
 }
 
@@ -85,5 +86,6 @@
   title = {Screencast: OMPI OpenFabrics Protocols},
   year = {2008},
   url = {https://www.open-mpi.org/video/openfabrics/ompi-openfabrics-protocols-v1.2-2up.pdf},
+  doi = {10.5281/zenodo.3339238},
   publisher = {CISCO}
 }
