### 2019-07-18 #20 @rohanmclure
I have the DOI links submitted now, thanks.
