### 2019-07-19 #20 @jpsamaroo
Thanks for that! I'm doing my best to make time to finish up my review of the code functionality/performance and paper content; in the meantime, please check out the following file for my comments on the paper so far: [ArrayChannels-review](https://git.sr.ht/~jpsamaroo/ArrayChannels-review/tree/master/paper.md)
