### 2019-07-21 #20 @rohanmclure
Thanks Julian, appreciate your comments, and will submit another commit acknowledging them tonight.
