### 2019-07-22 eda7474
Initial feedback response
