+++ paper/paper.tex
@@ -108,9 +108,9 @@ communication of data between processes in a parallel computation. This
 methodology is particularly useful when there is no notion of shared
 memory between processes, as in the case of distributed computing. Julia
 implements message passing through its \texttt{Distributed.jl} module in
-the standard library. The two main primitives available to the user are
-remote calls, as well as \texttt{Future} objects. Processes may message
-one another by means of a remote procedure call, whereby parameters to
+the standard library. The two main primitives available to the user are the
+\texttt{remotecall}, as well as \texttt{Future} objects. Processes may message
+one another by means of a remote procedure call, whereby arguments to
 the remote call and other captured variables are communicated to the
 recipient process. A \texttt{Future} fulfils the synchronising role of
 a remote call, encapsulating the completion state of function in
@@ -166,7 +166,7 @@ occur at a particular time (namely, following a successful handshake),
 the receiving process may offer in their intention to receive a preferred
 buffer which may be written to directly after synchronisation. In this
 way, rendezvous message passing may be used to avoid an extraneous
-memory copy, at the cost of additional synchronisation. While frameworks such as `Distributed.jl` that utilise eager message passing may omit the memory copy step, this mode will reduced temporal locality, as new buffers must be selected between communication events.
+memory copy, at the cost of additional synchronisation. While frameworks such as \texttt{Distributed.jl} that utilise eager message passing may omit the memory copy step, this mode will reduced temporal locality, as new buffers must be selected between communication events.
 
 \begin{figure}[htb]
 	\includegraphics[width=\linewidth]{figs/Rendezvous.pdf}
@@ -218,14 +218,14 @@ provided so as to provide each core with some local computation.
 
 \begin{figure}[htb]
 	\includegraphics[width=\linewidth]{figs/Reduce.pdf}
-	\caption{Reduce kernel}
+	\caption{Distributed reduce kernel}
 	\label{fig:reduce-diagram}
 \end{figure}
 
 In a distributed context, this is accomplished by allocating two large
 vectors at each locale, computing the local sum and then performing a
-distributed reduction on the resultant vector. In MPI, this is
-facilitated by directly by the \texttt{MPI\_Reduce} directive, which
+distributed reduction on the resultant vector. In figure~\ref{fig:reduce-diagram}, reduction occurs in stages occuring at different compute nodes for improved parallelism. In MPI, this is
+facilitated directly by the \texttt{MPI\_Reduce} directive, which
 causes the MPI runtime to conduct the parallel reduction with a single
 destination rank using whichever topology it deems most suitable. As the
 MPI Reduce implementation performs its work in place, our equivalent
@@ -242,10 +242,11 @@ representation of a dense matrix so that columns are stored in the row
 format and vice versa. In a distributed context, where the matrix is
 fragmented among different locales, the transpose of a pair of indices
 may belong to a different locale, and as such data communication is
-required. In practise, the matrices are distributed into `column
-blocks', with one column block given to each process. Figure~\ref{fig:transpose-diagram} demonstrates how off-diagonal regions of
+required. In practise, the matrices are distributed into "column
+blocks", with one column block given to each process. Figure~\ref{fig:transpose-diagram} demonstrates how off-diagonal regions of
[…]
