### 2019-07-22 #20 @rohanmclure
Hi Julian, I've responded to your recommendations so far. Thanks!
