+++ .travis.yml
@@ -1,6 +1,5 @@
 language: julia
 julia:
     - 1.1
-    - 1.3
 codecov: true
 
