+++ .travis.yml
@@ -1,5 +1,6 @@
 language: julia
 julia:
     - 1.1
+    - 1.3
 codecov: true
 
