### 2019-08-02 #20 @jpsamaroo
Ok, I've ticked off all the boxes, although I still have a few nitpicks here-and-there for both the paper and the code. I'll point these out with issues on the ArrayChannels.jl repository as I encounter them. Great work with this paper and the resulting code, I look forward to using it :smile:

@vchuravy
