+++ example/ping_pong_array_channels.jl
@@ -1,6 +1,5 @@
 using Distributed
 addprocs(2)
-@everywhere include("preload.jl")
 @everywhere using ArrayChannels
 
 function main()
