### 2020-07-27 #20 @vchuravy
@rohanmclure thank you for your patience, can you fix the DOIs? And is there an archival DOI for the associated code that you would like to associate with the publication?
https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance
