+++ paper/ref.bib
@@ -16,7 +16,7 @@
   title = {Structured Parallel Programming: Patterns for Efficient Computation},
   subtitle = {Patterns for efficient computation},
   url = {https://doi.org/10.1016/C2011-0-04251-5},
-  doi = {10.1016/C2011-004251-5},
+  doi = {10.1016/C2011-0-04251-5},
   year = {2012},
   publisher = {Morgan Kaufman}
 }
